reorganises module structure
